backend/recommendation_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
sentence-transformers>=2.2.2
huggingface_hub>=0.19.0
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.0
transformers>=4.30.0